Final release!

+++ sprintBacklog/models.py
@@ -5,7 +5,7 @@ from productBacklog.models import ProductBacklogItem
 
 class Task(models.Model):
     owner=models.ForeignKey(User, on_delete=models.CASCADE)
-    effort=models.IntegerField(default=0)
+    effort=models.PositiveSmallIntegerField(default=0)
     status=models.CharField(max_length=2, choices=[('TD','To-Do'),('P','In Progress'),('D','Done')],default='TD')
     pbi=models.ForeignKey(ProductBacklogItem, on_delete=models.CASCADE, blank=True, null=True)
     description=models.TextField(max_length=1000)
